+++ idea/testData/refactoring/rename/ambiguousClassFunImportRenameClass/after/lib/lib.kt
@@ -1,0 +1,5 @@
+package lib
+
+class Foo2
+
+fun Foo(p: Int): Foo2 = Foo2()