+++ idea/testData/refactoring/rename/clashWithInnerClass/after/test/test.kt
@@ -1,0 +1,9 @@
+package test
+
+class A {
+    inner class C {
+        inner class C {
+            val c: A.C = this@A.C()
+        }
+    }
+}