+++ idea/testData/refactoring/rename/clashWithNestedClass/before/test/test.kt
@@ -1,0 +1,9 @@
+package test
+
+class A {
+    class /*rename*/B {
+        class C {
+            val b: B = B()
+        }
+    }
+}